+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
 
+### Breaking changes
+
+- All previously deprecated functions have been removed (their replacements are still available).
+
 ## [0.4.0] - 2020-09-15
 
 ### Deprecations